+++ RibbonsColor/solution/visiblecolor/IVisibleColorFinder.cs
@@ -1,15 +1,10 @@
 ﻿using RibbonsColor.model;
-using System;
-using System.Collections.Generic;
 using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace RibbonsColor.solution.visiblecolor
 {
     public interface IVisibleColorFinder
     {
-        public Color FindVisibleColor(int pos, List<IRibbonModel> ribbons, ref int level);
+        public Color FindVisibleColor(long pos, List<IRibbonModel> ribbons, ref int level);
     }
 }
